pyproject.toml
[project]
name = "research-assistant-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.86.0",
    "arxiv>=2.4.1",
    "asyncio-throttle>=1.0.2",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.11.4",
    "fastapi>=0.135.2",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "kafka-python>=2.3.0",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.21",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "llama-cloud>=1.6.0",
    "llama-index>=0.14.18",
    "llama-index-indices-managed-llama-cloud>=0.11.0",
    "lxml>=6.0.2",
    "markdown>=3.10.2",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "semanticscholar>=0.11.0",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "streamlit-extras>=1.0.0",
    "tavily-python>=0.7.23",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.42.0",
]
